pyproject.toml
[project]
name = "quackbot"
version = "0.1.0"
description = "Manage House Group Chat"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.6",
]

[tool.pyright]
include = ["./"]
venvPath = ".venv"

# These are checked by ruff
reportUndefinedVariable = false
reportUnusedVariable = false
reportUnusedImport = false

[tool.ruff]
line-length = 130

[tool.ruff.lint]
extend-fixable = ["I", "ANN201"]
extend-select = ["ALL"]
ignore = [
    "ANN",
    "D1",
    "E501",
    "ERA001",
    "G010",
    "INP001",
    "PGH003",
    "PLR2004",
    "PLW2901",
    "PTH123",
    "PTH208",
    "S311",
    "SIM105",
    "TD",
    "TID",
    "UP015",
    "RUF001",
]
